Move transferapp functionality out of apps
