+++ nbgrader/apps/collectapp.py
@@ -1,35 +1,26 @@
-import os
-import glob
-import shutil
-from collections import defaultdict
+from traitlets import default
 
-from traitlets import Bool
-
-from .baseapp import TransferApp, transfer_aliases, transfer_flags
-from ..utils import check_mode, parse_utc
+from .baseapp import NbGrader, nbgrader_aliases, nbgrader_flags
+from ..exchange import Exchange, ExchangeCollect, ExchangeError
 
 
 aliases = {}
-aliases.update(transfer_aliases)
+aliases.update(nbgrader_aliases)
 aliases.update({
+    "timezone": "Exchange.timezone",
+    "course": "Exchange.course_id",
 })
 
 flags = {}
-flags.update(transfer_flags)
+flags.update(nbgrader_flags)
 flags.update({
     'update': (
-        {'CollectApp' : {'update': True}},
+        {'ExchangeCollect' : {'update': True}},
         "Update existing submissions with ones that have newer timestamps."
     ),
 })
 
-def groupby(l, key=lambda x: x):
-    d = defaultdict(list)
-    for item in l:
-        d[key(item)].append(item)
-    return d
-
-class CollectApp(TransferApp):
+class CollectApp(NbGrader):
 
     name = u'nbgrader-collect'
     description = u'Collect an assignment from the nbgrader exchange'
@@ -68,75 +59,39 @@ class CollectApp(TransferApp):
             nbgrader collect --update assignment1
         """
 
-    update = Bool(
-        False,
-        help="Update existing submissions with ones that have newer timestamps."
-    ).tag(config=True)
+    @default("classes")
+    def _classes_default(self):
+        classes = super(CollectApp, self)._classes_default()
+        classes.extend([Exchange, ExchangeCollect])
+        return classes
 
-    def _path_to_record(self, path):
-        filename = os.path.split(path)[1]
-        # Only split twice on +, giving three components. This allows usernames with +.
-        filename_list = filename.rsplit('+', 2)
-        if len(filename_list) != 3:
-            self.fail("Invalid filename: {}".format(filename))
-        username = filename_list[0]
-        timestamp = parse_utc(filename_list[2])
-        return {'username': username, 'filename': filename, 'timestamp': timestamp}
+    def _load_config(self, cfg, **kwargs):
+        if 'CollectApp' in cfg:
+            self.log.warn(
+                "Use ExchangeCollect in config, not CollectApp. Outdated config:\n%s",
+                '\n'.join(
+                    'CollectApp.{key} = {value!r}'.format(key=key, value=value)
+                    for key, value in cfg.CollectApp.items()
+                )
+            )
+            cfg.ExchangeCollect.merge(cfg.CollectApp)
+            del cfg.CollectApp
 
-    def _sort_by_timestamp(self, records):
-        return sorted(records, key=lambda item: item['timestamp'], reverse=True)
+        super(CollectApp, self)._load_config(cfg, **kwargs)
 
-    def init_src(self):
-        if self.exchange.course_id == '':
-            self.fail("No course id specified. Re-run with --course flag.")
+    def start(self):
+        super(CollectApp, self).start()
 
-        self.course_path = os.path.join(self.exchange.root, self.exchange.course_id)
-        self.inbound_path = os.path.join(self.course_path, 'inbound')
-        if not os.path.isdir(self.inbound_path):
-            self.fail("Course not found: {}".format(self.inbound_path))
-        if not check_mode(self.inbound_path, read=True, execute=True):
-            self.fail("You don't have read permissions for the directory: {}".format(self.inbound_path))
-        student_id = self.coursedir.student_id if self.coursedir.student_id else '*'
-        pattern = os.path.join(self.inbound_path, '{}+{}+*'.format(student_id, self.coursedir.assignment_id))
-        records = [self._path_to_record(f) for f in glob.glob(pattern)]
-        usergroups = groupby(records, lambda item: item['username'])
-        self.src_records = [self._sort_by_timestamp(v)[0] for v in usergroups.values()]
+        # set assignemnt and course
+        if len(self.extra_args) == 1:
+            self.coursedir.assignment_id = self.extra_args[0]
+        elif len(self.extra_args) > 2:
+            self.fail("Too many arguments")
+        elif self.coursedir.assignment_id == "":
+            self.fail("Must provide assignment name:\nnbgrader <command> ASSIGNMENT [ --course COURSE ]")
 
-    def init_dest(self):
-        pass
-
-    def copy_files(self):
-        for rec in self.src_records:
-            student_id = rec['username']
-            src_path = os.path.join(self.inbound_path, rec['filename'])
-            dest_path = self.coursedir.format_path(self.coursedir.submitted_directory, student_id, self.coursedir.assignment_id)
-            if not os.path.exists(os.path.dirname(dest_path)):
-                os.makedirs(os.path.dirname(dest_path))
-
-            copy = False
-            updating = False
-            if os.path.isdir(dest_path):
-                existing_timestamp = self._get_existing_timestamp(dest_path)
-                new_timestamp = rec['timestamp']
-                if self.update and (existing_timestamp is None or new_timestamp > existing_timestamp):
-                    copy = True
-                    updating = True
-            else:
-                copy = True
-
-            if copy:
-                if updating:
-                    self.log.info("Updating submission: {} {}".format(student_id, self.coursedir.assignment_id))
-                    shutil.rmtree(dest_path)
-                else:
-                    self.log.info("Collecting submission: {} {}".format(student_id, self.coursedir.assignment_id))
-                self.do_copy(src_path, dest_path)
-            else:
-                if self.update:
-                    self.log.info("No newer submission to collect: {} {}".format(
-                        student_id, self.coursedir.assignment_id
-                    ))
-                else:
-                    self.log.info("Submission already exists, use --update to update: {} {}".format(
-                        student_id, self.coursedir.assignment_id
-                    ))
+        collect = ExchangeCollect(coursedir=self.coursedir, parent=self)
+        try:
+            collect.start()
+        except ExchangeError:
+            self.fail("nbgrader collect failed")
