+++ nbgrader/apps/fetchapp.py
@@ -1,28 +1,27 @@
-import os
-import shutil
+from traitlets import default
 
-from textwrap import dedent
-from traitlets import Bool
-
-from .baseapp import TransferApp, transfer_aliases, transfer_flags
-from ..utils import check_mode
+from .baseapp import NbGrader, nbgrader_aliases, nbgrader_flags
+from ..exchange import Exchange, ExchangeFetch, ExchangeError
 
 
 aliases = {}
-aliases.update(transfer_aliases)
+aliases.update(nbgrader_aliases)
 aliases.update({
+    "timezone": "Exchange.timezone",
+    "course": "Exchange.course_id",
 })
 
 flags = {}
-flags.update(transfer_flags)
+flags.update(nbgrader_flags)
 flags.update({
     'replace': (
-        {'FetchApp' : {'replace_missing_files' : True}},
+        {'ExchangeFetch' : {'replace_missing_files' : True}},
         "replace missing files, even if the assignment has already been fetched"
     ),
 })
 
-class FetchApp(TransferApp):
+
+class FetchApp(NbGrader):
 
     name = u'nbgrader-fetch'
     description = u'Fetch an assignment from the nbgrader exchange'
@@ -51,66 +50,39 @@ class FetchApp(TransferApp):
         to turn in the assignment.
         """
 
-    replace_missing_files = Bool(False, help="Whether to replace missing files on fetch").tag(config=True)
+    @default("classes")
+    def _classes_default(self):
+        classes = super(FetchApp, self)._classes_default()
+        classes.extend([Exchange, ExchangeFetch])
+        return classes
 
-    def init_src(self):
-        if self.exchange.course_id == '':
-            self.fail("No course id specified. Re-run with --course flag.")
+    def _load_config(self, cfg, **kwargs):
+        if 'FetchApp' in cfg:
+            self.log.warn(
+                "Use ExchangeFetch in config, not FetchApp. Outdated config:\n%s",
+                '\n'.join(
+                    'FetchApp.{key} = {value!r}'.format(key=key, value=value)
+                    for key, value in cfg.FetchApp.items()
+                )
+            )
+            cfg.ExchangeFetch.merge(cfg.FetchApp)
+            del cfg.FetchApp
 
-        self.course_path = os.path.join(self.exchange.root, self.exchange.course_id)
-        self.outbound_path = os.path.join(self.course_path, 'outbound')
-        self.src_path = os.path.join(self.outbound_path, self.coursedir.assignment_id)
-        if not os.path.isdir(self.src_path):
-            self.fail("Assignment not found: {}".format(self.src_path))
-        if not check_mode(self.src_path, read=True, execute=True):
-            self.fail("You don't have read permissions for the directory: {}".format(self.src_path))
+        super(FetchApp, self)._load_config(cfg, **kwargs)
 
-    def init_dest(self):
-        if self.exchange.path_includes_course:
-            root = os.path.join(self.exchange.course_id, self.coursedir.assignment_id)
-        else:
-            root = self.coursedir.assignment_id
-        self.dest_path = os.path.abspath(os.path.join('.', root))
-        if os.path.isdir(self.dest_path) and not self.replace_missing_files:
-            self.fail("You already have a copy of the assignment in this directory: {}".format(root))
+    def start(self):
+        super(FetchApp, self).start()
 
-    def copy_if_missing(self, src, dest, ignore=None):
-        filenames = sorted(os.listdir(src))
-        if ignore:
-            bad_filenames = ignore(src, filenames)
-            filenames = sorted(list(set(filenames) - bad_filenames))
+        # set assignemnt and course
+        if len(self.extra_args) == 1:
+            self.coursedir.assignment_id = self.extra_args[0]
+        elif len(self.extra_args) > 2:
+            self.fail("Too many arguments")
+        elif self.coursedir.assignment_id == "":
+            self.fail("Must provide assignment name:\nnbgrader <command> ASSIGNMENT [ --course COURSE ]")
 
-        for filename in filenames:
-            srcpath = os.path.join(src, filename)
-            destpath = os.path.join(dest, filename)
-            relpath = os.path.relpath(destpath, os.getcwd())
-            if not os.path.exists(destpath):
-                if os.path.isdir(srcpath):
-                    self.log.warn("Creating missing directory '%s'", relpath)
-                    os.mkdir(destpath)
-
-                else:
-                    self.log.warn("Replacing missing file '%s'", relpath)
-                    shutil.copy(srcpath, destpath)
-
-            if os.path.isdir(srcpath):
-                self.copy_if_missing(srcpath, destpath, ignore=ignore)
-
-
-    def do_copy(self, src, dest, perms=None):
-        """Copy the src dir to the dest dir omitting the self.coursedir.ignore globs."""
-        if os.path.isdir(self.dest_path):
-            self.copy_if_missing(src, dest, ignore=shutil.ignore_patterns(*self.coursedir.ignore))
-        else:
-            shutil.copytree(src, dest, ignore=shutil.ignore_patterns(*self.coursedir.ignore))
-
-        if perms:
-            for dirname, dirnames, filenames in os.walk(dest):
-                for filename in filenames:
-                    os.chmod(os.path.join(dirname, filename), perms)
-
-    def copy_files(self):
-        self.log.info("Source: {}".format(self.src_path))
-        self.log.info("Destination: {}".format(self.dest_path))
-        self.do_copy(self.src_path, self.dest_path)
-        self.log.info("Fetched as: {} {}".format(self.exchange.course_id, self.coursedir.assignment_id))
+        fetch = ExchangeFetch(coursedir=self.coursedir, parent=self)
+        try:
+            fetch.start()
+        except ExchangeError:
+            self.fail("nbgrader fetch failed")
