+++ nbgrader/apps/listapp.py
@@ -1,41 +1,38 @@
-import os
-import glob
-import shutil
-import re
-import json
+from traitlets import default
 
-from traitlets import Bool
-
-from .baseapp import TransferApp, transfer_aliases, transfer_flags
+from .baseapp import NbGrader, nbgrader_aliases, nbgrader_flags
+from ..exchange import Exchange, ExchangeList, ExchangeError
 
 
 aliases = {}
-aliases.update(transfer_aliases)
+aliases.update(nbgrader_aliases)
 aliases.update({
+    "timezone": "Exchange.timezone",
+    "course": "Exchange.course_id",
 })
 
 flags = {}
-flags.update(transfer_flags)
+flags.update(nbgrader_flags)
 flags.update({
     'inbound': (
-        {'ListApp' : {'inbound': True}},
+        {'ExchangeList' : {'inbound': True}},
         "List inbound files rather than outbound."
     ),
     'cached': (
-        {'ListApp' : {'cached': True}},
+        {'ExchangeList' : {'cached': True}},
         "List cached files rather than inbound/outbound."
     ),
     'remove': (
-        {'ListApp' : {'remove': True}},
+        {'ExchangeList' : {'remove': True}},
         "Remove an assignment from the exchange."
     ),
     'json': (
-        {'ListApp' : {'as_json': True}},
+        {'ExchangeList' : {'as_json': True}},
         "Print out assignments as json."
     ),
 })
 
-class ListApp(TransferApp):
+class ListApp(NbGrader):
 
     name = u'nbgrader-list'
     description = u'List assignments in the nbgrader exchange'
@@ -89,133 +86,37 @@ class ListApp(TransferApp):
             nbgrader list --inbound --remove --student=student1
         """
 
-    inbound = Bool(False, help="List inbound files rather than outbound.").tag(config=True)
-    cached = Bool(False, help="List assignments in submission cache.").tag(config=True)
-    remove = Bool(False, help="Remove, rather than list files.").tag(config=True)
-    as_json = Bool(False, help="Print out assignments as json").tag(config=True)
+    @default("classes")
+    def _classes_default(self):
+        classes = super(ListApp, self)._classes_default()
+        classes.extend([Exchange, ExchangeList])
+        return classes
 
-    def init_src(self):
-        pass
+    def _load_config(self, cfg, **kwargs):
+        if 'ListApp' in cfg:
+            self.log.warn(
+                "Use ExchangeList in config, not ListApp. Outdated config:\n%s",
+                '\n'.join(
+                    'ListApp.{key} = {value!r}'.format(key=key, value=value)
+                    for key, value in cfg.ListApp.items()
+                )
+            )
+            cfg.ExchangeList.merge(cfg.ListApp)
+            del cfg.ListApp
 
-    def init_dest(self):
-        course_id = self.exchange.course_id if self.exchange.course_id else '*'
-        assignment_id = self.coursedir.assignment_id if self.coursedir.assignment_id else '*'
-        student_id = self.coursedir.student_id if self.coursedir.student_id else '*'
-
-        if self.inbound:
-            pattern = os.path.join(self.exchange.root, course_id, 'inbound', '{}+{}+*'.format(student_id, assignment_id))
-        elif self.cached:
-            pattern = os.path.join(self.exchange.cache, course_id, '{}+{}+*'.format(student_id, assignment_id))
-        else:
-            pattern = os.path.join(self.exchange.root, course_id, 'outbound', '{}'.format(assignment_id))
-
-        self.assignments = sorted(glob.glob(pattern))
-
-    def parse_assignment(self, assignment):
-        if self.inbound:
-            regexp = r".*/(?P<course_id>.*)/inbound/(?P<student_id>.*)\+(?P<assignment_id>.*)\+(?P<timestamp>.*)"
-        elif self.cached:
-            regexp = r".*/(?P<course_id>.*)/(?P<student_id>.*)\+(?P<assignment_id>.*)\+(?P<timestamp>.*)"
-        else:
-            regexp = r".*/(?P<course_id>.*)/outbound/(?P<assignment_id>.*)"
-
-        m = re.match(regexp, assignment)
-        if m is None:
-            raise RuntimeError("Could not match '%s' with regexp '%s'", assignment, regexp)
-        return m.groupdict()
-
-    def format_inbound_assignment(self, info):
-        return "{course_id} {student_id} {assignment_id} {timestamp}".format(**info)
-
-    def format_outbound_assignment(self, info):
-        msg = "{course_id} {assignment_id}".format(**info)
-        if os.path.exists(info['assignment_id']):
-            msg += " (already downloaded)"
-        return msg
-
-    def copy_files(self):
-        pass
-
-    def parse_assignments(self):
-        assignments = []
-        for path in self.assignments:
-            info = self.parse_assignment(path)
-            if self.exchange.path_includes_course:
-                root = os.path.join(info['course_id'], info['assignment_id'])
-            else:
-                root = info['assignment_id']
-
-            if self.inbound or self.cached:
-                info['status'] = 'submitted'
-                info['path'] = path
-            elif os.path.exists(root):
-                info['status'] = 'fetched'
-                info['path'] = os.path.abspath(root)
-            else:
-                info['status'] = 'released'
-                info['path'] = path
-
-            if self.remove:
-                info['status'] = 'removed'
-
-            info['notebooks'] = []
-            for notebook in sorted(glob.glob(os.path.join(info['path'], '*.ipynb'))):
-                info['notebooks'].append({
-                    'notebook_id': os.path.splitext(os.path.split(notebook)[1])[0],
-                    'path': os.path.abspath(notebook)
-                })
-
-            assignments.append(info)
-
-        return assignments
-
-    def list_files(self):
-        """List files."""
-        assignments = self.parse_assignments()
-
-        if self.as_json:
-            print(json.dumps(assignments))
-
-        else:
-            if self.inbound or self.cached:
-                self.log.info("Submitted assignments:")
-                for info in assignments:
-                    self.log.info(self.format_inbound_assignment(info))
-            else:
-                self.log.info("Released assignments:")
-                for info in assignments:
-                    self.log.info(self.format_outbound_assignment(info))
-
-    def remove_files(self):
-        """List and remove files."""
-        assignments = self.parse_assignments()
-
-        if self.as_json:
-            print(json.dumps(assignments))
-
-        else:
-            if self.inbound or self.cached:
-                self.log.info("Removing submitted assignments:")
-                for info in assignments:
-                    self.log.info(self.format_inbound_assignment(info))
-            else:
-                self.log.info("Removing released assignments:")
-                for info in assignments:
-                    self.log.info(self.format_outbound_assignment(info))
-
-        for assignment in self.assignments:
-            shutil.rmtree(assignment)
+        super(ListApp, self)._load_config(cfg, **kwargs)
 
     def start(self):
-        if self.inbound and self.cached:
-            self.fail("Options --inbound and --cached are incompatible.")
-
-        if len(self.extra_args) == 0:
-            self.extra_args = ["*"] # allow user to not put in assignment
-
         super(ListApp, self).start()
 
-        if self.remove:
-            self.remove_files()
-        else:
-            self.list_files()
+        # set assignemnt and course
+        if len(self.extra_args) == 1:
+            self.coursedir.assignment_id = self.extra_args[0]
+        elif len(self.extra_args) > 2:
+            self.fail("Too many arguments")
+
+        lister = ExchangeList(coursedir=self.coursedir, parent=self)
+        try:
+            lister.start()
+        except ExchangeError:
+            self.fail("nbgrader list failed")
