+++ nbgrader/exchange/exchange.py
@@ -1,14 +1,21 @@
 import os
 import datetime
+import sys
+import shutil
 
 from textwrap import dedent
 
 from dateutil.tz import gettz
 from traitlets.config import LoggingConfigurable
-from traitlets import Unicode, Bool, default
+from traitlets import Unicode, Bool, Instance, default
 from jupyter_core.paths import jupyter_data_dir
 
-from .utils import check_directory
+from ..utils import check_directory
+from ..coursedir import CourseDirectory
+
+
+class ExchangeError(Exception):
+    pass
 
 
 class Exchange(LoggingConfigurable):
@@ -60,6 +67,18 @@ class Exchange(LoggingConfigurable):
         )
     ).tag(config=True)
 
+    coursedir = Instance(CourseDirectory, allow_none=True)
+
+    def __init__(self, coursedir=None, **kwargs):
+        self.coursedir = coursedir
+        super(Exchange, self).__init__(**kwargs)
+        if self.parent:
+            self.log = self.parent.log
+
+    def fail(self, msg):
+        self.log.fatal(msg)
+        raise ExchangeError(msg)
+
     def set_timestamp(self):
         """Set the timestap using the configured timezone."""
         tz = gettz(self.timezone)
@@ -81,3 +100,32 @@ class Exchange(LoggingConfigurable):
         """See if the exchange directory exists and is writable, fail if not."""
         if not check_directory(self.root, write=True, execute=True):
             self.fail("Unwritable directory, please contact your instructor: {}".format(self.root))
+
+    def init_src(self):
+        """Compute and check the source paths for the transfer."""
+        raise NotImplementedError
+
+    def init_dest(self):
+        """Compute and check the destination paths for the transfer."""
+        raise NotImplementedError
+
+    def copy_files(self):
+        """Actually do the file transfer."""
+        raise NotImplementedError
+
+    def do_copy(self, src, dest, perms=None):
+        """Copy the src dir to the dest dir omitting the self.coursedir.ignore globs."""
+        shutil.copytree(src, dest, ignore=shutil.ignore_patterns(*self.coursedir.ignore))
+        if perms:
+            self.set_perms(dest, perms=perms)
+
+    def start(self):
+        if sys.platform == 'win32':
+            self.fail("Sorry, the exchange is not available on Windows.")
+
+        self.ensure_root()
+        self.set_timestamp()
+
+        self.init_src()
+        self.init_dest()
+        self.copy_files()
